requirements.txt
pyautogui
sounddevice
soundfile
whisper
pynput
python-dotenv
